Codechange: Move GUI parts of badges to a separate file.
