+++ src/CMakeLists.txt
@@ -279,6 +279,8 @@ add_files(
     newgrf_animation_type.h
     newgrf_badge.cpp
     newgrf_badge.h
+    newgrf_badge_gui.cpp
+    newgrf_badge_gui.h
     newgrf_badge_type.h
     newgrf_callbacks.h
     newgrf_canal.cpp
