+++ src/rail_gui.cpp
@@ -18,6 +18,7 @@
 #include "command_func.h"
 #include "waypoint_func.h"
 #include "newgrf_badge.h"
+#include "newgrf_badge_gui.h"
 #include "newgrf_station.h"
 #include "company_base.h"
 #include "strings_func.h"
